% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirmVars.R
\name{confirmVars}
\alias{confirmVars}
\title{Biomarker confirmation}
\usage{
confirmVars(
  data = NULL,
  standardization = FALSE,
  columnWise = TRUE,
  biomks = NULL,
  allmks = FALSE,
  outcomeType = c("binary", "continuous", "time-to-event"),
  Y = NULL,
  time = NULL,
  event = NULL,
  outfile = "nameWithPath"
)
}
\arguments{
\item{data}{A data matrix or a data frame, samples are in columns, and features/traits including outcome and biomarkers are in rows.}

\item{standardization}{A logic parameter to indicate if standardization is needed before biomarker confirmation/validation, the default value is FALSE.}

\item{columnWise}{A logic variable to indicate if column wise or row wise normalization is needed for standardization process, the default is TRUE. 
This is only meaningful when "standardization" is TRUE.}

\item{biomks}{A vector of biomarker names to be confirmed/validated, they should be a subset of "data" column names. Each of these biomarkers will be 
confirmed/validated separately, and all together if allmks is TRUE.}

\item{allmks}{A logic variable to indicate if all "biomks" should be analyzed together in addition to single variable analysis. The default is FALSE.}

\item{outcomeType}{Outcome variable type. There are three choices: "binary" (default), "continuous", and "time-to-event".}

\item{Y}{Outcome varialbe name when the outcome type is either "binary" or "continuous".}

\item{time}{Time variable name when outcome type is "time-to-event".}

\item{event}{Event variable name when outcome type is "time-to-event".}

\item{outfile}{A string for output files including path if necessary but without file type extension.}
}
\description{
This is the wrap up function to confirm/validate known biomarkers in a given data set
}
\details{
This function is to confirm/validate if a single variable or a group of variables has or have effect on an outcome variable for a given data set.
An outcome variable could be a binary, continuous, or a time-to-event variable.
}
\references{

}
\author{
Aixiang Jiang
}
\keyword{biomaker}
\keyword{biomarker}
\keyword{confirmation,}
\keyword{validation}
