% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datlist}
\alias{datlist}
\title{This is an example data in csmpv}
\format{
A list with training and a validation datasets.
}
\usage{
datlist
}
\description{
This dataset contains sample data for csmpv.
}
\author{
Aixiang Jiang
}
\keyword{datasets}
