% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO2_predict.R
\name{LASSO2_predict}
\alias{LASSO2_predict}
\title{Predict and Validate LASSO2 Model Scores}
\usage{
LASSO2_predict(
  lassoObj,
  newdata = NULL,
  newY = FALSE,
  u = 2,
  outfile = "nameWithPath"
)
}
\arguments{
\item{lassoObj}{A LASSO2 object.}

\item{newdata}{A new data matrix or data frame where samples are in rows and features/traits are in columns.}

\item{newY}{The outcome variable for the new data set.}

\item{u}{A single numeric follow-up time for survival outcomes.}

\item{outfile}{A string for the output file including the path, if necessary, but without the file type extension.}
}
\value{
A vector of predicted values is returned.
}
\description{
This function predicts model scores, positive group probabilities, and risk scores for different outcome types based on a LASSO2 object and a new data set.
It also performs validation of predictions when the true outcome variable is available.
}
\references{
Friedman, J., Hastie, T. and Tibshirani, R. (2008) Regularization Paths for Generalized Linear Models via Coordinate Descent (2010), Journal of Statistical Software, Vol. 33(1), 1-22, doi:10.18637/jss.v033.i01.
  Simon, N., Friedman, J., Hastie, T. and Tibshirani, R. (2011) Regularization Paths for Cox's Proportional Hazards Model via Coordinate Descent, Journal of Statistical Software, Vol. 39(5), 1-13, doi:10.18637/jss.v039.i05.
  Harrell Jr F (2023). rms: Regression Modeling Strategies_. R package version 6.7-1, <https://CRAN.R-project.org/package=rms>
  Harrell Jr F (2023). Hmisc: Harrell Miscellaneous_. R package version 5.1-1, <https://CRAN.R-project.org/package=Hmisc>
}
\author{
Aixiang Jiang
}
