% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LASSO_plus.R
\name{LASSO_plus}
\alias{LASSO_plus}
\title{LASSO_plus is for variable selection, and model buildup with the most commonly used R functions: lm, glm, and coxph for different outcome type.}
\usage{
LASSO_plus(
  data = NULL,
  standardization = FALSE,
  columnWise = TRUE,
  biomks = NULL,
  outcomeType = c("binary", "continuous", "time-to-event"),
  Y = NULL,
  time = NULL,
  event = NULL,
  topN = 10,
  outfile = "nameWithPath",
  height = 6
)
}
\arguments{
\item{data}{A data matrix or a data frame, samples are in columns, and features/traits are in rows.}

\item{standardization}{A logic variable to indicate if standardization is needed before variable selection, the default is FALSE.}

\item{columnWise}{A logic variable to indicate if column wise or row wise normalization is needed, the default is TRUE, which us tod do column-wise normalization. 
This is only meaningful when "standardization" is TRUE.}

\item{biomks}{A vector of potential biomarkers for variable selection, they should be a subset of "data" column names.}

\item{outcomeType}{Outcome variable type. There are three choices: "binary" (default), "continuous", and "time-to-event".}

\item{Y}{Outcome varialbe name when the outcome type is either "binary" or "continuous".}

\item{time}{Time variable name when outcome type is "time-to-event".}

\item{event}{Event variable name when outcome type is "time-to-event".}

\item{topN}{An interger to indicate how many variables we intend to select}

\item{outfile}{A string for the output file including path if necessary but without file type extension.}

\item{height}{An integer to indicate the forest plot height in inches}
}
\value{
A model is returned: 
\item{fit}{a model with selected variables for the given outcome variable}
}
\description{
This is the wrap up function to select variables that are associated with an outcome variable based on my new algorithm: LASSO_plus, 
       and build up a model afterwards.
}
\details{
This function is to use LASSO_plus algorithm to select variables that are assoicated with on an outcome variable for a given data set.
An outcome variable could be a binary, continuous, or time-to-event variable. LASSO_plus selects variables based on LASSO, single variable regession, 
and stepwise regression. After variable selection, a model is built-up as well.
}
\references{

}
\author{
Aixiang Jiang
}
\keyword{selection}
\keyword{variable}
