% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGpred.R
\name{XGpred}
\alias{XGpred}
\title{XGpred: Building Risk Classification Predictive Models using Survival Data}
\usage{
XGpred(
  data = NULL,
  varsIn = NULL,
  time = NULL,
  event = NULL,
  nrounds = 5,
  probcut = 0.8,
  outfile = "nameWithPath"
)
}
\arguments{
\item{data}{A data matrix or a data frame, samples are in rows, and features/traits are in columns}

\item{varsIn}{A vector of variables used for prediction model}

\item{time}{Time variable name}

\item{event}{Event variable name}

\item{nrounds}{An integer to indicate how many times}

\item{probcut}{Probability cutoff for risk group classification}

\item{outfile}{A string for the output file including path if necessary but without file type extension.}
}
\value{
A list is returned with the following seven items:
\item{XGBplus}{XGBplus object}
\item{stables}{Stabel high and low risk group samples identified by XGBplus}
\item{weights}{Weights for each variables used in the model}
\item{modelPars}{Mean and standard error of model scores for each risk group}
\item{XGpred_score}{Model XGpred score}
\item{XGpred_prob}{Empirical Bayesian probability based on model XGpred score}
\item{XGpred_prob_class}{Risk group classification based on XGpred_prob for the given probability cutoff}
\item{probcut}{Probability cutoff for risk group classification}
}
\description{
The XGpred function serves as a wrapper for our novel XGpred algorithm. This algorithm combines XGBoost, clustering, and survival probability prediction
to create stable high and low-risk groups based on survival data. These groups are subsequently filtered and leveraged for constructing two models: an XGpred linear prediction score model and an empirical Bayesian-based binary risk classification model.
}
\details{
It's important to note that this function does not include a variable selection step. All provided variables will be used in constructing the predictive models.
If variable selection is desired, the LASSO_plus function within the csmpv R package can be called.
}
\references{
Tianqi Chen and Carlos Guestrin, "XGBoost: A Scalable Tree Boosting System", 22nd SIGKDD Conference on Knowledge Discovery and Data Mining, 2016, https://arxiv.org/abs/1603.02754
  Aoki T, Jiang A, Xu A et al. The spatially resolved tumor microenvironment predicts treatment outcome in relapsed/refractory Hodgkin lymphoma,
         bioRxiv 2023.05.19.541331; doi: https://doi.org/10.1101/2023.05.19.541331
}
\author{
Aixiang Jiang
}
