% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XGpred_predict.R
\name{XGpred_predict}
\alias{XGpred_predict}
\title{Predicting Risk Group Classification for a New Data Set}
\usage{
XGpred_predict(newdat = NULL, XGpredObj = NULL, scoreShift = 0)
}
\arguments{
\item{newdat}{A data matrix or a data frame where samples are in rows and features/traits are in columns. It should include all variables needed
for the prediction model.}

\item{XGpredObj}{An XGpred object returned by the XGpred function. Although not all items are needed for prediction purposes, using the XGpred object as input is convenient.}

\item{scoreShift}{A calibration value to subtract from the current model score if needed.}
}
\value{
A data frame containing XGpred_score, XGpred_prob, and XGpred_prob_class
}
\description{
The XGpred.predict function is designed to predict risk group classifications for a new data set. This prediction is based on the assumption that an XGpred object is available from a training data set and that the new data set is comparable to the training data set.

In scenarios where the new and training data sets are not directly comparable, a calibration cohort is required to ensure accurate risk group predictions.
}
\references{
Aoki T, Jiang A, Xu A et al. The spatially resolved tumor microenvironment predicts treatment outcome in relapsed/refractory Hodgkin lymphoma,
         bioRxiv 2023.05.19.541331; doi: https://doi.org/10.1101/2023.05.19.541331
}
\author{
Aixiang Jiang
}
